+++ .gitignore
@@ -14,3 +14,5 @@ hs_err_pid*
 .classpath
 .project
 /bin
+/target/
+.settings
