+++ src/main/resources/typIFC2X3_TC1.ser
